pyproject.toml
[project]
name = "website-environment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langdetect>=1.0.9",
    "mysql-connector-python>=9.4.0",
    "ollama>=0.5.4",
    "openai>=1.108.1",
    "pandas>=2.3.2",
    "pydantic>=2.11.9",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tara"]
